+++ integreat_cms/static/src/js/pois/poi-actions.ts
@@ -15,6 +15,9 @@ window.addEventListener("load", () => {
       // Reset to initial value
       longitude.value = longitude.dataset.initial;
       latitude.value = latitude.dataset.initial;
+
+      longitude.dispatchEvent(new Event("focusout"));
+      latitude.dispatchEvent(new Event("focusout"));
     }
   });
 
@@ -79,12 +82,18 @@ async function autoCompleteAddress() {
     updateField("longitude", data.longitude);
     updateField("latitude", data.latitude);
   }
+
+  let longitude = document.getElementById("id_longitude") as HTMLInputElement;
+  let latitude = document.getElementById("id_latitude") as HTMLInputElement;
+  
+  longitude.dispatchEvent(new Event("focusout"));
+  latitude.dispatchEvent(new Event("focusout"));
 }
 
 function updateField(fieldName: string, value: string) {
   let field = document.getElementById(`id_${fieldName}`) as HTMLInputElement;
   // Only fill value if it was changed
-  if (field.value != value) {
+  if (value && field.value != value) {
     field.value = value;
     field.classList.add("!border-green-500");
     // Reset green border after 5 seconds